move TestUtils file into `testing` subpackage
